feat(optimizated): update pages and components
run eslint and build
