+++ src/app/globals.css
@@ -99,8 +99,8 @@ a {
   text-decoration: none;
 }
 
-@media (prefers-color-scheme: dark) {
+@media (prefers-color-scheme) {
   html {
-    color-scheme: dark;
+    color-scheme: light;
   }
 }
