+++ src/components/ui/sidenav/SideNavItems.ts
@@ -24,65 +24,65 @@ export function getNavItems(): SidenavItem[] {
   const liquidationMenu: SidenavMenuItem[] = [
     {
       label: 'Envío de Insumos',
-      to: '/dashboard/export/add-supply-shipment',
+      to: '/dashboard/liquidation/add-supply-shipment',
     },
     {
       label: 'Envíos Realizados',
-      to: '/liquidation/exports-sent',
+      to: '/dashboard/liquidation/exports-sent',
     },
     {
       label: 'Pago a Productores',
-      to: '/dashboard/export/producer-pending-payments',
+      to: '/dashboard/liquidation/producer-pending-payments',
     },
     {
       label: 'Pagos Realizados',
-      to: '/liquidation/producer-payments',
+      to: '/dashboard/liquidation/producer-payments',
     },
   ];
 
   const clientMenu: SidenavMenuItem[] = [
     {
-      label: 'Consultar Puerto',
-      to: '/dashboard/client/harbors',
-    },
-    {
       label: 'Agregar Puerto',
       to: '/dashboard/client/add-harbor',
     },
     {
-      label: 'Consultar Cliente',
-      to: '/dashboard/client/clients',
+      label: 'Consultar Puerto',
+      to: '/dashboard/client/harbors',
     },
     {
       label: 'Agregar Cliente',
       to: '/dashboard/client/add-client',
     },
     {
-      label: 'Cuentas Bancarias',
-      to: '/dashboard/client/bank-accounts',
+      label: 'Consultar Cliente',
+      to: '/dashboard/client/clients',
     },
     {
       label: 'Agregar Cuenta Bancaria',
       to: '/dashboard/client/add-bank-account',
     },
+    {
+      label: 'Cuentas Bancarias',
+      to: '/dashboard/client/bank-accounts',
+    },
   ];
 
   const productorMenu: SidenavMenuItem[] = [
-    {
-      label: 'Consultar Productor',
-      to: '/dashboard/producer/producers',
-    },
     {
       label: 'Agregar Productor',
       to: '/dashboard/producer/add-producer',
     },
     {
-      label: 'Consultar Finca',
-      to: '/dashboard/producer/fincas',
+      label: 'Consultar Productor',
+      to: '/dashboard/producer/producers',
     },
     {
       label: 'Agregar Finca',
       to: '/dashboard/producer/add-fincas',
+    },
+    {
+      label: 'Consultar Finca',
+      to: '/dashboard/producer/fincas',
     },
     {
       label: 'Agregar Logo',
@@ -96,35 +96,35 @@ export function getNavItems(): SidenavItem[] {
 
   const boxBrandsMenu: SidenavMenuItem[] = [
     {
-      label: 'Consultar Marca',
-      to: '/dashboard/box-brands/search',
-    },
-    {
       label: 'Agregar Marca de Caja',
       to: '/dashboard/box-brands/add-box-brand',
+    },
+    {
+      label: 'Consultar Marca',
+      to: '/dashboard/box-brands/search',
     },
     {
       label: 'Agregar Logo',
       to: '/dashboard/box-brands/upload-logo',
     },
     {
-      label: 'Tipos de Corte',
-      to: '/dashboard/box-brands/cutting-types',
-    },
-    {
       label: 'Agregar Tipo de Corte',
       to: '/dashboard/box-brands/add-cutting-type',
+    },
+    {
+      label: 'Tipos de Corte',
+      to: '/dashboard/box-brands/cutting-types',
     },
   ];
 
   const cuttingSheetsMenu: SidenavMenuItem[] = [
     {
-      label: 'Hojas de Corte',
-      to: '/dashboard/export/cutting-sheets',
-    },
-    {
       label: 'Agregar Hoja de Corte',
       to: '/dashboard/export/add-cutting-sheet',
+    },
+    {
+      label: 'Hojas de Corte',
+      to: '/dashboard/export/cutting-sheets',
     },
   ];
